+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+## v2.34.1 - 08 Aug 2023
+
 ### Fixes
 * [artifact-manager] 148 / vra-ng:pull command failure when import/export tags contains underscore
 * [maven-plugins] 146 / Pull operations fail on Windows with release 2.34.0